add optional ellipse to scatter+line chart, works similar to scatter+line
i needed to define an area around scattered points, i have data centered on a
point and i expect is to be within an ellipse of that point. data that falls
outside of this ellipse should be considered bad and checked upon.

this feature allows adding data and a ellipse center and radius to a group of
data. group would en up looking like:
  key: 'Group Name',
  values: [
    {x: someValueX, y: someValueY, size: someSize, shape: 'circle']}, ...
    ],
  // define these is ellipse is desired
  ellipseCx: someEllipseCxValue, // cx field of ellipse
  ellipseCy: someEllipseCyValue, // cy field of ellipse
  ellipseRx: someEllipseRxValue, // rx field of ellipse
  ellipseRy: someEllipseRyValue, // ry field of ellipse
  // define this only if regLine
  slope: Math.random() - .01,
  intercept: Math.random() - .5,

the ellipse used is the svg ellipse, as explained here:
  https://www.dashingd3js.com/svg-basic-shapes-and-d3js